alex/tailwind-exports-fix: split out
